Add spell checker project (#24)
## Changelist

Adds a spell checker project.

I expect this would be a solo project - at least as it stands - but
welcome feedback on extensions etc. I initially thought this might be a
bit bigger in scope, but many of the interesting aspects of this project
are in asking questions about their approach (e.g. "what might happen if
you replaced Basic English with real English?"), and I think those are
difficult to write a very precise rubric like I think we're aiming for
here. Having said that I still think this is a decent standalone
project.

---------

Co-authored-by: Sally McGrath <sally@codeyourfuture.io>
Co-authored-by: Daniel Wagner-Hall <daniel@codeyourfuture.io>