Clarify the Spaced Repetition Tracker project requirements (#47)
A few general clarifications of the Spaced Repetition Tracker spec,
based on some observations from the first cohort completing it:

- Clarify that revision dates in the past should not be shown in the
agenda
- Make the "user 3" instructions clear that the calculated time gaps are
the same for selecting dates in the future or in the past
- Remove the `getUserIds` scaffolding, as it was only implicit that it
should be used to generate the list of users. Removing it means that
trainees won't over-complicate things by joining the simple list of ids
with richer data (e.g. user names)
- Make it more explicit the user's agenda must be loaded from storage
when a user is selected. Hopefully prevent cases where they inject data
automatically
- Add form validation requirements. Most of the trainees in cohort 1 did
this anyway, but I think it's good to be explicit
- Clarify that no user should be auto-selected on page load